chore(TCOMP-317): move diToAvro() method to SchemaConverter
