+++ daikon/src/test/java/org/talend/daikon/di/DiIncomingSchemaEnforcerTest.java
@@ -579,290 +579,4 @@ public class DiIncomingSchemaEnforcerTest {
         assertNull(runtimeSchema);
     }
 
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns String avro schema in case "id_String" di type is
-     * passed
-     */
-    @Test
-    public void testDiToAvroString() {
-        Schema expectedSchema = AvroUtils._string();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_String", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns Boolean avro schema in case "id_Boolean" di type
-     * is passed
-     */
-    @Test
-    public void testDiToAvroBoolean() {
-        Schema expectedSchema = AvroUtils._boolean();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_Boolean", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns Integer avro schema in case "id_Integer" di type
-     * is passed
-     */
-    @Test
-    public void testDiToAvroInteger() {
-        Schema expectedSchema = AvroUtils._int();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_Integer", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns Long avro schema in case "id_Long" di type is
-     * passed
-     */
-    @Test
-    public void testDiToAvroLong() {
-        Schema expectedSchema = AvroUtils._long();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_Long", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns Double avro schema in case "id_Double" di type is
-     * passed
-     */
-    @Test
-    public void testDiToAvroDouble() {
-        Schema expectedSchema = AvroUtils._double();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_Double", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns Float avro schema in case "id_Float" di type is
-     * passed
-     */
-    @Test
-    public void testDiToAvroFloat() {
-        Schema expectedSchema = AvroUtils._float();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_Float", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns Integer avro schema with
-     * "java-class"=java.lang.Byte
-     * in case "id_Byte" di type is passed
-     */
-    @Test
-    public void testDiToAvroByte() {
-        Schema expectedSchema = AvroUtils._byte();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_Byte", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns Integer avro schema with
-     * "java-class"=java.lang.Short
-     * in case "id_Short" di type is passed
-     */
-    @Test
-    public void testDiToAvroShort() {
-        Schema expectedSchema = AvroUtils._short();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_Short", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns String avro schema with
-     * "java-class"=java.lang.Character
-     * in case "id_Character" di type is passed
-     */
-    @Test
-    public void testDiToAvroCharacter() {
-        Schema expectedSchema = AvroUtils._character();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_Character", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns String avro schema with
-     * "java-class"=java.math.BigDecimal
-     * in case "id_BigDecimal" di type is passed
-     */
-    @Test
-    public void testDiToAvroBigDecimal() {
-        Schema expectedSchema = AvroUtils._decimal();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_BigDecimal", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns Long avro schema with "java-class"=java.util.Date
-     * in case "id_Date" di type is passed
-     */
-    @Test
-    public void testDiToAvroDate() {
-        Schema expectedSchema = AvroUtils._date();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro("id_Date", null);
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} throws {@link UnsupportedOperationException}
-     * in case unsupported type is passed
-     */
-    @Test
-    public void testDiToAvroNotSupporter() {
-        thrown.expect(UnsupportedOperationException.class);
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        enforcer.diToAvro("Unsupported", null);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns logical date avro schema with in case "date"
-     * logical type is passed
-     */
-    @Test
-    public void testDiToAvroLogicalDate() {
-        Schema expectedSchema = AvroUtils._logicalDate();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro(null, "date");
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns logical time-millis avro schema with in case
-     * "time-millis" logical type is passed
-     */
-    @Test
-    public void testDiToAvroLogicalTime() {
-        Schema expectedSchema = AvroUtils._logicalTime();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro(null, "time-millis");
-
-        assertEquals(expectedSchema, actualSchema);
-    }
-
-    /**
-     * Checks {@link DiIncomingSchemaEnforcer#diToAvro(String, String)} returns logical timestamp-millis avro schema with in case
-     * "timestamp-millis" logical type is passed
-     */
-    @Test
-    public void testDiToAvroLogicalTimestamp() {
-        Schema expectedSchema = AvroUtils._logicalTimestamp();
-
-        Schema designSchema = SchemaBuilder.builder().record("Record") //
-                .fields() //
-                .name("booleanField").type().booleanType().noDefault() //
-                .endRecord();
-
-        DiIncomingSchemaEnforcer enforcer = new DiIncomingSchemaEnforcer(designSchema);
-        Schema actualSchema = enforcer.diToAvro(null, "timestamp-millis");
-
-        assertEquals(expectedSchema, actualSchema);
-    }
 }
